requirements.txt
appnope
asttokens
comm
debugpy
decorator
exceptiongroup
executing
filelock
fsspec
ipykernel
ipython
jedi
Jinja2
jupyter_client
jupyter_core
MarkupSafe
matplotlib-inline
mpmath
nest-asyncio
networkx
numpy
packaging
parso
pexpect
pillow
platformdirs
prompt_toolkit
psutil
ptyprocess
pure_eval
Pygments
python-dateutil
pyzmq
six
stack-data
sympy
torch
torchvision
tornado
tqdm
traitlets
typing_extensions
wcwidth
pyyaml
aiohttp
datasets
gdown
importlib-metadata
jaraco.collections
matplotlib
opencv-python
pysocks
tensorflow-hub
tomli